requirements.txt
fastapi
uvicorn
jinja2
ccxtpro
